refactor: BoardService, MemberService 수정
