+++ src/main/java/com/example/bcsd/entity/Article.java
@@ -42,6 +42,12 @@ public class Article {
         this.content = content;
     }
 
+    public void update(String title, String content) {
+        this.title = title;
+        this.content = content;
+        this.modifiedAt = LocalDateTime.now();
+    }
+
     public void setId(Long id) {
         this.id = id;
     }
