+++ src/main/java/com/example/bcsd/service/ArticleService.java
@@ -5,6 +5,9 @@ import com.example.bcsd.controller.dto.response.ArticlesInBoardViewResponse;
 import com.example.bcsd.controller.dto.request.ArticleCreateRequest;
 import com.example.bcsd.controller.dto.response.ArticleResponse;
 import com.example.bcsd.controller.dto.request.ArticleUpdateRequest;
+import com.example.bcsd.entity.Article;
+import com.example.bcsd.entity.Board;
+import com.example.bcsd.entity.Member;
 import com.example.bcsd.global.exception.CustomException;
 import com.example.bcsd.global.exception.ErrorCode;
 import com.example.bcsd.repository.*;
